collect workout from google fit
